Adding progress tracking for uploadTagPrefix cases
Summary:
Adding progress tracking, but for now only to the cases when uploadTagPrefix is used.
The reason we're not adding it to cases when user uses the uploadTag option for now is that there is a risk that they might upload a different file with the same uploadTag thus leading to us using progress tracking for the previous file and rejecting some data even when it can be new.

Test Plan: Try uploading the same file twice and verify that you get an error when uploading same rows the second time.

Reviewers: ytx, ryanher

Reviewed By: ytx, ryanher

Differential Revision: https://phabricator.intern.facebook.com/D5373457

Tasks: 19919360

Signature: t1:5373457:1499382224:2ce2d40cff61bbd892f62202fb89ae7f46566037